Add main menu, save slots, and redesigned hostile events
Main Menu & Save System:
- Added main menu screen shown on startup
- 3 save slots with evidence, ascensions, playtime display
- Continue, New Game, Delete save options per slot
- Quick Continue button for most recent save
- Return to menu button (☰) next to sound toggle

Redesigned Hostile Events:
- Evidence Thief: Now moves around screen, chase and click to catch
- Tinfoil Thief: Progress bar mechanic, click to push back
  - Scales with tinfoil amount (harder with more tinfoil)
- Debunker: Unchanged (rapid click challenge)

New Content:
- 15 new end-game generators (Tiers 2-4)
- 13 new quests (max 25 min duration)
- 18 new tinfoil shop items (Advanced & Elite tiers)
- New upgrades for post-ascension generators

Balance Changes:
- Reduced mid/late generator costs for ~1 hour first ascension
- BelieverBonus changed from int to double for large values

+++ Models/Generator.cs
@@ -8,7 +8,7 @@ public class Generator
     public required double BaseCost { get; init; }
     public double CostMultiplier { get; init; } = 1.15;
     public required double BaseProduction { get; init; }
-    public int BelieverBonus { get; init; }
+    public double BelieverBonus { get; init; }
 
     public double GetCost(int owned)
     {
